hash_image/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
ImageHash==4.3.2
numpy==2.3.5
pillow==12.0.0
psycopg==3.3.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyWavelets==1.9.0
scipy==1.16.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
